+++ crates/graphql-composition/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Improvements
+
+- Implemented support for the `@internal` directive from the composite schemas spec (https://github.com/grafbase/grafbase/pull/3185)
+
 ## 0.8.0 - 2025-05-16
 
 ### Breaking changes